chore: number is a top-level module
