+++ src/number/mod.rs
@@ -1,0 +1,5 @@
+mod number;
+mod promote;
+
+pub use number::Num;
+pub use promote::promote_to_array;
